+++ frontend/src/components/mainTitle/mainTitle.tsx
