+++ frontend/src/index.tsx
@@ -1,13 +1,11 @@
 import React, { FC } from 'react';
 import ReactDOM from 'react-dom';
 import { createGlobalStyle } from 'styled-components';
-import {
-  BrowserRouter, Link, Route, Routes,
-} from 'react-router-dom';
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import LoginPage from './pages/login/login';
 import Pixeboy from '../public/font/Pixeboy.ttf';
 import RegistrationPage from './pages/registration/registration';
-import MainLayout from './layouts/main-layout';
+import MainLayout from './layouts/mainLayout';
 
 const GS = createGlobalStyle`
   @font-face {
@@ -27,13 +25,10 @@ const App: FC = () => {
   return <React.Fragment>
     <BrowserRouter>
       <GS/>
-
       <Routes>
         <Route path='/login' element={<MainLayout><LoginPage/></MainLayout>}/>
         <Route path='/registration' element={<MainLayout><RegistrationPage/></MainLayout>}/>
       </Routes>
-      <Link to={'login'}>Login</Link>
-
     </BrowserRouter>
   </React.Fragment>;
 };
